odg/api/types: `ArtefactExtraID` is a `map[string]string`
According to the OCM spec it should be a `map[string]string`, although the
Delivery Service API will accept anything.

The issue with Delivery Service API accepting anything is that it causes
exceptions in the API, due to poor validation of user-provided payloads.

Related: https://github.com/gardener/cc-utils/pull/1204

Discussed with the CI/CD folks and they'll be working on improving the error
handling on their side, so that clients don't see 500 Internal Server Errors.